fix(backend): Poetry project organization + more
