+++ backend/classes/LandSatData.py
@@ -1,0 +1,2 @@
+class LandSatData:
+    raise NotImplementedError
