+++ backend/classes/__init__.py
@@ -1,0 +1,5 @@
+from backend.classes.LandSatAPI import LandSatAPI
+from backend.classes.LandSatData import LandSatData
+
+LandSatAPI = LandSatAPI
+LandSatData = LandSatData
