+++ backend/routers/__init__.py
@@ -1,0 +1,5 @@
+from backend.routers.coordinates_router import coordinates_router
+from backend.routers.index_router import index_router
+
+index_router = index_router
+coordinates_router = coordinates_router
