--- backend/src/classes/LandSatData.py
@@ -1,2 +1,0 @@
-class LandSatData:
-    raise NotImplementedError
