--- backend/src/classes/__init__.py
@@ -1,5 +1,0 @@
-from backend.src.classes.LandSatAPI import LandSatAPI
-from backend.src.classes.LandSatData import LandSatData
-
-LandSatAPI = LandSatAPI
-LandSatData = LandSatData
