--- backend/src/routers/__init__.py
@@ -1,5 +1,0 @@
-from backend.src.routers.index_router import index_router
-from backend.src.routers.coordinates_router import coordinates_router
-
-index_router = index_router
-coordinates_router = coordinates_router
