--- backend/src/types/__init__.py
@@ -1,3 +1,0 @@
-from typing import List, Tuple
-
-bbox_type = Tuple[int, int, int, int]
