+++ backend/types/__init__.py
@@ -1,0 +1,3 @@
+from typing import Tuple
+
+bbox_type = Tuple[int, int, int, int]
