+++ package.json
@@ -5,7 +5,9 @@
     "frontend:dev": "pnpm --prefix ./frontend/ dev",
     "frontend:lint": "pnpm --prefix ./frontend/ lint",
     "frontend:format": "pnpm --prefix ./frontend/ format",
-    "backend:dev": "cd backend && poetry run uvicorn src.main:app --reload",
+    "backend:dev": "poetry run uvicorn backend.main:app --reload",
+    "backend:lint": "poetry run ruff check",
+    "backend:format": "poetry run ruff check --fix",
     "dev": "pnpm run /:dev$/",
     "lint": "pnpm run /:lint$/",
     "format": "pnpm run /:format$/",
