+++ pyproject.toml
@@ -8,7 +8,7 @@ description = ""
 authors = ["EKK Ingarden"]
 license = "GPL-3.0"
 readme = "README.md"
-package-mode = false
+packages = [{include = "backend"}]
 
 [tool.poetry.dependencies]
 python = "^3.12"
@@ -30,6 +30,8 @@ select = [
     # isort
     "I001",
 ]
+
+src = ["backend"]
 
 [build-system]
 requires = ["poetry-core"]
